Prevent Schema::Loader from overwriting unchanged schema files
previously, the make_schema_at feature of Schema::Loader would overwrite
files in the dump_to_dir regardless of whether the underlying database
schema had changed or not. This patch causes Schema::Loader to check
the contents of the current file with the newly created schema from the
database. The file is only overwritten if the two differ. This means
that the old timestamp and creation date are preserved.