remove placeholders from tab/col comments queries
Changes the use placeholders in ::DBI _table_comment and _column_comment
methods to inline $dbh->quote protected values.

This is for DBD::Sybase with ASE versions that do not support
placeholders and MSSQL connections with a low tds version.

At some point we'll have methods for doing queries that do the right
thing depending on what the driver/database supports.