add -I option to dbicdump
Add an -I option to dbicdump to add things to @INC during the dump, like
perl's -I option (see perldoc perlrun.)

For the config file dbicdump format, add the 'lib' key that is the
equivalent of the -I option.

Add error checking to ::Utils slurp_file and write_file subs in the open
calls.

Add alnewkirk to list of CONTRIBUTORS for dbicdump.