skip dbicdump tests on Win32 due to test fails
t/23dumpmore.t and t/26dump_use_moose.t fail on Win32 with some types of
refs and characters  in -o params to dbicdump because of quoting issues.

There is currently no module like String::ShellQuote for Win32, haarg is
releasing Win32::ShellQuote soon, and then I will reenable these tests.

Also the fucking pigs completely fucked up my windows computer and when
I was trying to run some tests with dbicdump I got constant perl hangs,
so I cannot debug this right now and I disabled ALL dbicdump tests on
Win32.

I removed the reference to the FUCKING_PIGS environment variable
because, at least, the behavior of the tests is now at least consistent
with that of other machines, or was, except for the perl hangs I'm
getting now, and because I'm making this skip for all dbicdump tests,
not just those with refs so it is in the same part of the code, so might
as well.