make common tests use fewer active statements
Newer versions of the Sybase ASE developer's edition have a limit of 25
user connections and do not allow setting it higher, this did not
support the common tests.

Change the ->first calls on ResultSets to ->single to not keep active
statements around.

The common tests now run against the latest version of Sybase ASE
developer's edition.