introspect ON DELETE/UPDATE clauses for MySQL FKs
The default for these has been CASCADE for a while now, the idea was
that this is the most useful default for deployed schemas and it would
be replaced with the actual introspected value when this capability
would be added.

A user wrote in to RT that this capability was missing for MySQL so I
decided to start on adding it, with MySQL first.

This means that for MySQL the default for on_delete/on_update is now
RESTRICT rather than CASCADE, and this will probably soon be true for
other backends that support these clauses.

Also there are some whitespace changes in this commit due to some
cleanup autocommands I put into my .vimrc, read it with git diff -w.