introspect ON DELETE/UPDATE and DEFERRABLE for pg
Apparently Pg has an information_schema, so I copied the _table_fk_info
from the MSSQL driver and modified it to introspect the ON and
DEFERRABLE clauses.

I will add this introspection for MSSQL shortly as well, after which I
will remove the duplication from these two drivers and add an
InformationSchema component.