support coderef for relationship_attrs
Add support for passing a coderef as relationship_attrs and describe
thorougly in the ::Base POD (as well as improve the existing POD for
 it.)

This is the list of parameters:

 * rel_name

 * local_source

 * remote_source

 * local_table

 * local_cols

 * remote_table

 * remote_cols

 * attrs

The hash is passed as a list, not as a hashref.

The coderef must return a hashref of attributes or nothing.

This is thoroughly tested in t/45relationships.t.

I had to add a fallback => 1 to the DBObject.pm stringification
overload, so that eq would work, and this makes the stringification much
more functional.