introspect ON DELETE/UPDATE for SQL Anywhere
The default seems to be RESTRICT, but it's hard to tell, because the
actions are implemented as triggers and you have to query the catalog
triggers view to get at them. At least if you specify RESTRICT there is
no trigger generated.

SQL Anywhere does not support DEFERRABLE constraints. It has something
called CHECK ON COMMIT but that does not seem to have the same full
functionality, so for now we leave is_deferrable at 1.