Skip many_to_many bridges involving might_have relationships
If one or both of the relationsips is a one-to-one relatiohip,
the bridge should be one-to-many or one-to-one, respectively.

However, DBIx::Class doesn't support such bridges, so skip them
for now, keeping the option open to add support if/when DBIC does,
without having to worry about back-compat shenanigans.

This reverts commit b0cdc8a5bc6db9206a53e52bd9c868ba8442ca2b
and fixes the crash differently.