Set $result_source_class->table_class appropriately on views
This constitutes partial support for the per-backend TODO items labelled
"introspect view SQL".  I've therefore left those items unchanged for the
backends that are now known to support this, and added a note for this part
of the work to the other backends.

Backends whose DBD implements enough of the DBI table_info method to indicate
whether a relation is in fact a view should also acquire support for setting
table_class, but there are no tests for other backends, so they haven't been
declared working.