Remove pointless disconnect after static ->load
It was added on the assumption that in static mode nothing else was
going to happen after the loading was done, but that's really none of
our concern, so we shouldn't interfere with it.

It also broke the in-transaction dump test with DBIC's new stricter
connectedness checking on txn ops (dbsrgits/dbix-class@d4702d53).