Use {List,Sub}::Util instead of List::MoreUtils and Sub::Name
That means bumping the dependency on List::Util to 1.33, which means
people on perl < 5.20 will have to upgrade anyway, so we might as well
use Sub::Util (added in 1.40) instead of Sub::Name as well.

Nick the apply, firstidx and uniq functions from List::MoreUtils::PP
into DBIx::Class::Schema::Loader::Utils to avoid depending on it.