Tweak "exclude" and "constraint" option docs
A user reading the documentation online may click on a TOC link to the
"exclude" option, not seeing that the description also applies to the
"constraint" option immediately above. This commit seeks to remedy that
by moving the description to the "constraint" section and giving
"exclude" and explicit backward reference to it.