do not require MooseX::MarkAsMethods with only_autoclean=1
When use_moose=1 and only_autoclean=1, we do not require
MooseX::MarkAsMethods, but require a more recent version of Moose. Check
for whichever set of prereqs is appropriate.

Technically, we don't need MXMAM at all if Moose is at 2.1400, but the
generated code will still (for now) use the MXMAM constructs if
only_autoclean is not set.  This gives the user more flexibility if
their Moose installation isn't at a new enough version on all systems.