Handle current_timestamp() from MariaDB DESCRIBE
At some point, MariaDB started outputing 'current_timestamp()' in the
default field when DESCRIBE is called on the table. This is a change
from 'CURRENT_TIMESTAMP' in mysql and older versions of mariadb.  As
such, our equality match started to fail and resulting schema dumps
produced `current_timestamp()` instead of `\"current_timestamp"`